agents/accounts-receivable-clerk/requirements.txt
# Core LangGraph and OpenAI
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
openai>=1.0.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Direct API integrations (available in braid core)
# These are imported from core.integrations, no additional installs needed

# Optional: Enhanced document processing
# Uncomment if you want advanced PDF/document parsing
# pymupdf>=1.24.0
# python-docx>=1.1.0

# Communication tools for live demo
twilio>=8.0.0
slack-sdk>=3.0.0

# Google API for Gmail and Drive integration
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0

# Optional: Enhanced scheduling
# Uncomment for production scheduling features
# apscheduler>=3.10.0
# celery>=5.3.0